feat(mobileApp): convert parseOrganizations to TS
